Move DebugClear RPC to a separate Debug gRPC service
