+++ cmd/ateapi/internal/debugapi/debug_clear.go
@@ -1,0 +1,36 @@
+// Copyright 2026 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package debugapi
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/agent-substrate/substrate/pkg/proto/ateapipb"
+)
+
+func (s *Service) DebugClear(ctx context.Context, req *ateapipb.DebugClearRequest) (*ateapipb.DebugClearResponse, error) {
+	if err := validateDebugClearRequest(req); err != nil {
+		return nil, err
+	}
+	if err := s.persistence.DebugClearAll(ctx); err != nil {
+		return nil, fmt.Errorf("while running DebugClearAll: %w", err)
+	}
+	return &ateapipb.DebugClearResponse{}, nil
+}
+
+func validateDebugClearRequest(req *ateapipb.DebugClearRequest) error {
+	return nil
+}
